backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
pandas==2.1.3
geopy==2.4.1
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
email-validator==2.1.0
